+++ gpflow/conditionals/multioutput/__init__.py
@@ -1,0 +1,2 @@
+from . import sample_conditionals
+from . import conditionals
