+++ gpflow/covariances/multioutput/__init__.py
@@ -1,0 +1,1 @@
+from . import kufs, kuus
