+++ gpflow/inducing_variables/multioutput/__init__.py
@@ -1,7 +1,4 @@
-from .inducing_variables import InducingVariables, InducingPoints, Multiscale
-from .inducing_patch import InducingPatches
-from . import multioutput
-from .multioutput import (
+from .inducing_variables import (
     MultioutputInducingVariables,
     FallbackSharedIndependentInducingVariables,
     FallbackSeparateIndependentInducingVariables,
