+++ gpflow/kernels/multioutput/__init__.py
@@ -1,0 +1,7 @@
+from .kernels import (
+    MultioutputKernel,
+    SeparateIndependent,
+    SharedIndependent,
+    IndependentLatent,
+    LinearCoregionalization,
+)
